Reapply "Merge pull request #29 from NachiketKandari/db-migration"
This reverts commit 5142f4d843eb86370aec3248a868015dac0acd8e.

+++ backend/src/insightxpert/training/schema.py
@@ -6,7 +6,7 @@ CREATE TABLE transactions (
     timestamp TEXT,
     transaction_type TEXT,
     merchant_category TEXT,
-    amount_inr REAL,
+    amount_inr DOUBLE PRECISION,
     transaction_status TEXT,
     sender_age_group TEXT,
     receiver_age_group TEXT,
